chore: separated body parsing tests
